// frontend/src/app/api-gen/models/execute-sql-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { StatementResultInfo } from '../models/statement-result-info';
export interface ExecuteSqlResponse {
  env: string;
  err_count: number;
  object_type: string;
  ok_count: number;
  results?: Array<StatementResultInfo>;
  schema_name: string;
}